Aula 180 - Upload - Atualizar Produto - Imagens - Parte 7

+++ LojaVirtual/LojaVirtual/Libraries/Arquivo/GerenciadorArquivo.cs
@@ -55,22 +55,27 @@ namespace LojaVirtual.Libraries.Arquivo
                 {
                     var NomeArquivo = Path.GetFileName(CaminhoTemp);
 
-                    var CaminhoAbsolutoTemp = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads/temp", NomeArquivo);
-                    var CaminhoAbsolutoDef = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads", ProdutoId.ToString(), NomeArquivo);
+                    var CaminhoDef = Path.Combine("uploads", ProdutoId.ToString(), NomeArquivo).Replace("\\", "/");
 
-                    if (File.Exists(CaminhoAbsolutoTemp))
+                    if (CaminhoDef != CaminhoTemp)
                     {
-                        File.Copy(CaminhoAbsolutoTemp, CaminhoAbsolutoDef);
-                        if (File.Exists(CaminhoAbsolutoDef))
+                        var CaminhoAbsolutoTemp = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads/temp", NomeArquivo);
+                        var CaminhoAbsolutoDef = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads", ProdutoId.ToString(), NomeArquivo);
+
+                        if (File.Exists(CaminhoAbsolutoTemp))
                         {
-                            File.Delete(CaminhoAbsolutoTemp);
-                        }
+                            File.Copy(CaminhoAbsolutoTemp, CaminhoAbsolutoDef);
+                            if (File.Exists(CaminhoAbsolutoDef))
+                            {
+                                File.Delete(CaminhoAbsolutoTemp);
+                            }
 
-                        ListaImagensDef.Add(new Imagem() { Caminho = Path.Combine("uploads", ProdutoId.ToString(), NomeArquivo).Replace("\\", "/"), ProdutoId = ProdutoId });
-                    }
-                    else
-                    {
-                        return null;
+                            ListaImagensDef.Add(new Imagem() { Caminho = Path.Combine("uploads", ProdutoId.ToString(), NomeArquivo).Replace("\\", "/"), ProdutoId = ProdutoId });
+                        }
+                        else
+                        {
+                            return null;
+                        }
                     }
 
                 }
